refact: `k-preview-sizes`
